+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+## v3.0.89(NOV 22, 2018)
+ * Added `createDistinctChannelIfNotExist()` in `GroupChannel`.
+   * It creates distinct channel and gets the channel with `isCreated` flag which represents whether the channel is actually created or not.
+ * Added `getMessageChangeLogsByTimestamp()` in `SendBird`.
+   * It retrieves message change logs since the given timestamp.
+ * Added `hiddenState` property in `GroupChannel` to get channel's hidden state.
+   * `unhidden`: It's not hidden channel.
+   * `hidden-allow-auto-unhide`: It's hidden channel which is automatically unhidden when new message comes in.
+   * `hidden-prevent-auto-unhide`: It's hidden channel which is not unhidden when new message comes in.
+ * Added `allowAutoUnhide` as the second parameter into `hide()` to allow automatic unhide on new message.
+ * Added `unhide()` in `GroupChannel` to manually unhide the channel.
+ * Added `hiddenChannelFilter` in `GroupChannelListQuery` to filter by `hiddenState`.
+   * `unhidden_only`: Get all unhidden channels. (default)
+   * `hidden_only`: Get all hidden channels which `hiddenState` is `hidden_allow_auto_unhide` or `hidden_prevent_auto_unhide`.
+   * `hidden_allow_auto_unhide`: Get channels which `hiddenState` is `hidden_allow_auto_unhide`.
+   * `hidden_prevent_auto_unhide`: Get channels which `hiddenState` is `hidden_prevent_auto_unhide`.
+
 ## v3.0.88(NOV 15, 2018)
  * Changed type of `sender` property in `UserMessage` and `FileMessage` from `User` to a new class `Sender` which extends `User`.
    * `Sender` has `isBlockedByMe` property which indicates that the message sender is blocked by the current user (default: false).